requirements.txt
fastapi
uvicorn
python-multipart
Pillow
pytesseract
pdf2image
openai
python-dotenv
pydantic
opencv-python
PyMuPDF